backend/requirements.txt
flask==3.1.2
flask-cors>=4.0.0
gunicorn>=20.1.0
firebase-admin==7.1.0
google-cloud-firestore==2.22.0
python-dotenv==1.2.2